Restructure to src-layout with private nanobind _core extension and abi3 wheels
